WIP: Partial fix for players being kicked for changing team, skin, or color. Still requires all the calls that are missing to setinfo(). Updated defs.qc to have built-ins that exist in PQWSV.

+++ Progs/make.sh
@@ -3,6 +3,11 @@ PATH_TO_FORTRESS_DIRECTORY="../../fortress"
 
 ./cppreqcc
 ./cpqccx /O2
-#pause
-mv ../prozac.dat $PATH_TO_FORTRESS_DIRECTORY/qwprozac.dat
+
+# If we compiled successfully, prozac.dat will have been put in "..".
+if test -f "../prozac.dat"; then
+    mv ../prozac.dat $PATH_TO_FORTRESS_DIRECTORY/qwprozac.dat
+    cp progdefs.h ../PQWSV/src/progdefs.h
+fi
+
 ./clean.sh
